Added Day/Night Cycle
Known issues:
Stars are on a sphere of radius 1000, rotating around 0,0,0. We may
eventually need to make the sphere and stars bigger and, or make the
heavenly bodies rotate around the player's position so that the sky
doesn't become distorted as the player reaches the edge of the map.
Stars are visible during the day and under the horizon.