Rain Effect and Generic Animal Behavior
I've been experimenting with a rain effect to make the weather system
more fleshed out.  It will cost to much to simulate rain everywhere on
the map. Instead, I'll have to create a raincloud that follows the
player's field of vision.

I also worked with Jami to create a Generic Animal behavior, complete
with Wandering and Meandering behaviors. It still has to be tested.