Rudimentary Camera Bag System
Basic workings of the camera bag and its interface.

Dragging items should work although there are no draggable items at this
point

To populate our camera bag with items we will probably need to create a
json database of all the items an then build at runtime.

The size of the camera bag in the interface took more hand authoring
than expected but is easily scalable for the most part.