Control GUI, Changing Lenses and Shutter Speed
and a little house cleaning.

Control gui is just some text in the upper left corner with the most
current controls.

We can now swap between our three lens types and I made prefabs for each
one so we can just throw the models on when they are made.  Only issue
is that the model doesn't place in front of the camera model. I'm
working on it...

Opted to replace our previous motion blur script with the far more
relevantly named camera motion blur and then I made a little function to
allow the player to change the "shutter speed" and therefore effect how
much blur they get.

Also I move things around a bit as the main asset folder was a jumble